[#28758] YSQL: reduce catalog cache memory spike on new connections
Summary:
During a stress test that involves a connection pool, it was found that sometimes new physical
connections come in a batch. Because the stress test executes a global-impact DDL frequently, it can
happen that all of the new connections in the batch will try to rebuild the relcache init file at
once. This can also happen during a cluster rolling restart. After a tserver is restarted, new
connections may also come concurrently. Because the relcache init file is deleted on a restart, all
the new connections will also try to rebuild the relcache init file. In the end, they all write
identical relcache init file but the side effect of doing relcache init file rebuild is that each
such connection will incur a memory spike due to preloading a set of catalog tables. By default, a
new connection that does not need to rebuild relcache file will incur much less catalog cache memory
cost. In addition, while preloading the set of catalog tables, the peak memory cost is twice of the
stable memory cost. The stress test sometimes see OOM error and this can happen when many new
connections end up rebuilding relcache init file unnecessarily.

To address this problem, this diff introduced a new PG to tserver RPC TriggerRelcacheInitConnection.
In particular, a connection that is not from the superuser "postgres" will not rebuild the relcache
init file. Instead, it will make a call TriggerRelcacheInitConnection to the local tserver. The local
tserver will trigger an internal connection as superuser "postgres" to rebuild the relcache init
file. The original connection will wait until the call TriggerRelcacheInitConnection returns, and
then resume its rest of the connection process by loading the newly generated relcache init file. In
this way we turn a new connection that would otherwise rebuilding relcache init file into one that
behaves as a "subsequent" connection, which does not need to rebuild relcache init file.

Note that if the original connection was made as the superuser "postgres", then we will not trigger
another internal connection. It is expected that in reality, direct connecting as "postgres" should
be rare.

On a per-database basis, when there are N concurrent connections to the same DB all need to rebuild
relcache init file, then all of them will send TriggerRelcacheInitConnection to the local tserver.
Only one of them will become the leader that makes a real internal connection to complete the
relcache init file rebuilding work. After that, all the requests will be satisfied and the RPC
TriggerRelcacheInitConnection is returned back to the waiting PG backend.

Note that the internal connection is short-lived because its only purpose is to rebuild the relcache
init file. Once the internal connection terminates, its spike of memory is released. In contrast,
those N connections would have stayed to execute queries for an extended period of time (determined
by connection pool policy on connection lifetime) with catalog cache memory footprint at an elevated
level. That makes the node more vulnerable to OOM error.

I plan to also investigate to use a real internal superuser instead of `postgres` for this purpose
because sometimes we do make direct connection as `postgres` and it will be better we can ensure
in the default mode (when preloading isn't enabled), relcache init is only performed by the
designated internal superuser connection.

One side effect of this diff is that in some unit tests, a connection used to be the first connection
to a given DB will no longer be the first connection due to the internally created connection. As
a result, the connection no longer performs relcache init file rebuild and preload a set of catalog
tables. I have to update 2 unit tests as a result of this. Also note that in such a case there could
be a performance side effect as well because before the diff we would have preloaded a lot more
catalog tables with the memory spike but it does have a good benefit to reduce first query latency.
After the diff we no longer preload that many catalog tables, so the first query will incur more
catalog cache misses which require on-demand master RPC for each of them.

**Upgrade/Rollback safety:**
 The src/yb/tserver/pg_client.proto change is only used in PG -> tserver communication which is upgrade safe.
Jira: DB-18472

Test Plan:
./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionTest.ConcurrentNonSuperuserNewConnectionsTest -n 100 --tp 2
./yb_build.sh release --java-test 'org.yb.pgsql.TestPgCostModelSeekNextEstimation'
YB_EXTRA_DAEMON_FLAGS="--enable_ysql_conn_mgr=true" ybd release --java-test 'org.yb.pgsql.TestPgCostModelSeekNextEstimation'

Reviewers: kfranz, sanketh, #db-approvers

Reviewed By: kfranz

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D47028