[#28699] DocDB: Avoid restoring to a time prior to PITR schedule max retention
Summary:
The diff 34068043b47e caused a regression where if the user wants to restore to a time earlier than the maximum retention window, i.e. `now - snapshot schedule retention`, we will silently try to create a new snapshot (as we didn't find a suitable snapshot that covers the restore_at time). We then use this newly created snapshot with a very old restore_at time, which leaves the database in a bad state.

Instead, we should have returned an error to the user that the `restore_to` time is earlier than the retention window.
The diff fixes the issue by checking if the `restore_to` time is earlier than the minimum allowed restore time and returns `IllegalState` error to the end user.

Test Plan:
./yb_build.sh --cxx-test yb-admin-snapshot-schedule-test --gtest_filter YbAdminSnapshotScheduleTest.FailRestoreToBeforeMaximumRetention

./yb_build.sh --cxx-test yb-admin-snapshot-schedule-test --gtest_filter YbAdminSnapshotScheduleTestLongPoll.FailRestoreWhenNoSnapshotCreated

Reviewers: asrivastava, zdrudi

Reviewed By: asrivastava, zdrudi

Subscribers: ybase, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D47054