[#28720] docdb: Initialize log segment size tracker correctly on segment reuse
Summary:
The `Log::active_segment_ondisk_size_` stores the current size of the active log segment file to enable adjusting the `wal_size` metric later on segment overflow or truncation. When setting the active log segment on log open, this variable needs to be set to the size of that log segment. It is being set correctly when the active log segment is a newly allocated segment, but not when an existing segment is reused.

This change sets the variable correctly when an existing segment is used as the active segment on log open and adds a test to verify that. As a precaution, I've set `active_segment_ondisk_size_` whenever `active_segment_sequence_number_` is set. The change also removes extraneous includes in `log_util.h` from my previous commit.
Jira: DB-18421

Test Plan:
```
./yb_build.sh --cxx-test='TEST_F(LogTest, TestLogMetricsWithSegmentReuse) {'
./yb_build.sh --cxx-test='TEST_F(LogTest, TestLogMetrics) {'
```
Jenkins

Reviewers: yyan, rthallam

Reviewed By: yyan

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D47052