[#28000] YSQL: Fix PgRecursiveAbortTest.AbortOnTserverFailure with txn ddl
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in autonomous transactions and commit independently from the overall transaction block it was part of.

This revision fixes the test `AbortOnTserverFailure` to work with Transactional DDL. The key difference lies in the fact that the DDL `create table t2` does not trigger the mocked `FinishTransaction` RPC call with txn ddl since it no longer auto-commits. The fix is to add a separate `COMMIT` call  to trigger the failure.
Jira: DB-17619

Test Plan:
Jenkins: test regex: .*AbortOnTserverFailure.*

Enabled transactional ddl manually and ran
```
./yb_build.sh --cxx-test pgwrapper_pg_mini-test --gtest_filter PgRecursiveAbortTest.AbortOnTserverFailure
```

Reviewers: kramanathan

Reviewed By: kramanathan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47063