[PLAT-18654] Filter export configuration options based on support
Summary:
This diff adds support for filtering the export configuration options based on the form context.
When asking users to choose an export configuration for db audit log export, we present only the export
configurations which support logs export.
When asking users to choose an export configuration for metrics export, we present only the export configurations
which support metrics export.

Test Plan:
Verify the options for db audit logs export and metrics export are filtered and present only the intended
options.
{F399110}
{F399111}

Reviewers: skurapati, rmadhavan

Reviewed By: skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D47120