[#28435] YSQL: Ensure read permission to group and others on initdb log file
Summary:
YBM requires that the group and others have read permission on the log files.

The permissions of postgresql-*.log are controlled by log_file_mode GUC. GUCs are not available during initdb. This revision hardcodes read permission for the group and others on the initdb log file. During runtime, initdb executes in local mode (IsYugaByteLocalNodeInitdb() is true), which only involves the bootstrap backend process. So this change is localised to the bootstrap backend process.
Jira: DB-18117

Test Plan:
Manual testing on a local universe.

Before this change
```
-rw-------  1 aagrawal  staff       0 Sep 29 21:18 initdb.log
```

With this change
```
-rw-r--r--  1 aagrawal  staff       0 Oct  1 20:26 initdb.log
```

Reviewers: kramanathan

Reviewed By: kramanathan

Subscribers: smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D47084