[BACKPORT 2025.1][#28623] DocDB: Add memory tracker for temporary memory used by reading WALs
Summary:
Original commit: d4747fb6f7a7be99858e261d8fd56c2420e128d2 / D46829
No conflicts.

Currently we have no tracking of the memory holding WAL records read from disk.  (The code that reads these records and allocates the memory is in LogCache.)

This can lead to a TServer crashing due to running out of memory – see #22992: DocDB Xcluster node becomes unreachable due to OOM..

Here we add a new memory tracker, Log Reader Memory, which tracks how much memory is currently being used to hold batches from LogReader.

The memory tracker is created in tablet_memory_manager.cc and passed down through
  * TSTabletManager::OpenTablet in yb/tserver/ts_tablet_manager.cc passes down TabletMemoryManager::read_wal_mem_tracker()
    * ditto SysCatalogTable::OpenTablet in yb/master/sys_catalog.cc
  * to BootstrapTablet in yb/tablet/tablet_bootstrap_if.cc
  * to BootstrapTabletImpl in yb/tablet/tablet_bootstrap.cc
  * to BootstrapTablet in yb/tablet/tablet_bootstrap.cc
  * to:
    * Log::Open -> Log::Log in yb/consensus/log.cc
      * -> LogReader::Open and ReadableLogSegment
    * LogReader::Open -> LogReader in yb/consensus/log_reader.cc
      * -> ReadableLogSegment
until it reaches ReadableLogSegment in yb/consensus/log_util.cc.

This tracker is passed as a shared_ptr and explicitly allowed to be nullptr in unit tests and tools for convenience.

There, ReadableLogSegment::ReadEntryBatch consumes memory as the batch is allocated; the shared_ptr it returns will call a destructor when the memory is no longer needed, which in turn will update the tracker with the release of that batch's memory.

Note there is currently no limit on this tracker; this will be added in the next diff.

Fixes #28623
Jira: DB-18318

Test Plan:
I am running a QA pipeline, https://perf.dev.yugabyte.com/perfstudio?pipelineId=90, which runs out of WAL memory, crashing the TServer.

the following chart shows the new tracker spiking as the TServer runs out of memory
{F395171}
 this doesn't capture all the memory increase as there are other temporary memory structures downstream of the WAL memory.
{F395172}
nonetheless, we have captured the majority (90%?),  Which should be sufficient for controlling this memory surge
in this test there were no adjustments so probably all the batches were under 4 KiB

For this diff I am just looking for the presence of this metric and that it correlates well with the (otherwise untracked) memory used.

running locally to verify tracker exists:
```
Memory usage by subsystem

 Id  Current  Peak consumption  Limit
   Consumption
 root  17.99M  18.06M  20.27G
 ...
 Log Reader Memory->server->root  0B  20.2K  none
```

Reviewers: xCluster, jhe, mhaddad

Reviewed By: jhe

Subscribers: ybase, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D47160