[#28766] DocDB: Fix a race between alter table and create snapshot
Summary:
Imagine the following scenario on a table `t(k int primary key)`:

1. `ALTER TABLE t ADD COLUMN v INT` command is issued to a pg backend. The command starts executing and is not yet committed in the following steps.
2. pg_backend writes the changes related to this DDL to the pg catalog as intents (not committed yet).
3. pg_backend sends an AlterTable RPC to the master leader.
4. Master leader executed the alter, which means saving the new schema of the DocDB table to the sys.catalog (this is committed) and creating a `YsqlDdlTxnVerifierState` object to track whether the transaction is committed or not from pg.
5. pg_backend commits the transaction, which means the writes to the pg catalog are now committed.
6. pg_backend notifies the master leader that the transaction is committed, which means the master will roll forward and clear the `YsqlDdlTxnVerifierState`.

If a create snapshot RPC arrives after step 4 but before step 5 and picks a `snapshot_hybrid_time` between 4 and 5, then the snapshot will contain the new DocDB schema for table t (that includes column `v`). However, running ysql_dump as of `snapshot_ht` will return the Create table t with the old schema (without column `v`). This is because the new column is not yet committed from pg side.

This leads to inconsistency between the ysql_dump output and the DocDB schema in the snapshot metadata. Consequently, the restore will fail as at restore, we are comparing the DocDB schema of the table exported from the backup side with the DocDB schema created by executing the ysql_dump. Restore was failing if these 2 schemas were not identical.

The root cause of the problem is that the DDL atomicity project guarantees "eventual consistency" between the pg and DocDB catalogs. This means there is a small window during the DDL execution where pg schema and the DocDB schemas are not consistent. That doesn't impose a correctness problem on the typical read/write path. However, as we want to support backups during DDLs, this inconsistency starts imposing a problem.

The diff fixes the issue by figuring out the committed DocDB schema on the restore side. The solution is based on the following invariant:
C1- The DocDB schema created when running the dump is the correct committed schema and can be used as the source of truth. This is because on the backup side, ysql_dump dumps the committed schema from pg perspective as of the desired `snapshot_ht`.

The changes in the ImportSnapshot workflow are as follows:
1- While iterating through the tables inside the `SnapshotInfoPB`, if DocDB schema for a table has `ysql_ddl_txn_verifier_state`, then skip schema equality checks for this table. This is because this table was undergoing a DDL while the snapshot was created at the backup side. Thus, we cannot tell for sure whether its current DocDB schema is the committed one or not. However, we guarantee that the DocDB table at the restore side has the correct committed schema as per C1.
2- Bump the schema version for this table to 1 + the schema version of `SysTableEntryPB` found in the `snapshotInfo`.

The new semantics are as follows: At the tserver, all schema packings coming from the snapshot will be used in tablet-meta at the restore side, and the last schema will have the correct committed schema created at the restore side as part of executing the SQL dump. The last schema is sent from the master to the tservers during importSnapshot (the master sends an AlterTable rpc to tservers).

For example, if the last schema version from the `SysTableEntryPB` of the backup is 8, then the tablet-meta of the snapshot will have one of the following 2 options (depending on whether the AlterTable RPC reached the tserver or not at backup creation time):
- schema:[k int, V Text], schema_version:8, old_schema_packings: [0,1,2,3,4,5,6,7].
- schema:[k int], schema_version:7, old_schema_packings: [0,1,2,3,4,5,6].

After restore, the final tablet meta will be one of the 2:
- schema:[k int], schema_version:9*, old_schema_packings: [0,1,2,3,4,5,6,7,8].
- schema:[k int], schema_version:9*, old_schema_packings: [0,1,2,3,4,5,6,7].

It is also worth mentioning that this solution might create a gap in schema packings on the restore side. In the previous example, some tablets' snapshots might have old schema packing until 7. We are ok with this gap as it doesn't affect correctness.

Test Plan:
A set of tests was added to test a backup taken in the middle of different ALTER TABLE DDLs.
The DDLs added in this diff: `ADD COLUMN`, `DROP COLUMN`, `RENAME COLUMN`

We also have 2 test instances for every DDL, one with colocation and one without colocation.
./yb_build.sh --cxx-test yb-backup-during-ddl-test --gtest_filter AlterTable/YBBackupDuringAlterTable.AddColumn/*
./yb_build.sh --cxx-test yb-backup-during-ddl-test --gtest_filter AlterTable/YBBackupDuringAlterTable.DropColumn/*
./yb_build.sh --cxx-test yb-backup-during-ddl-test --gtest_filter AlterTable/YBBackupDuringAlterTable.RenameColumn/*

Reviewers: zdrudi, asrivastava

Reviewed By: zdrudi, asrivastava

Subscribers: ybase, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D48331