[BACKPORT 2025.2][#28000][#28791] YSQL: Merge catalog modification aspects after nested ddls with txn ddl
Summary:
This revision adds the missing logic to merge catalog modification aspects as part of nested ddls when transactional ddl is enabled.

The issue is best described via the `CREATE TEMP TABLE` statement. Note that the creation of temp table has two major steps:
1. Cleanup of any existing temp tables from previous sessions
2. Creation of the new temp table

Key point: If the statement only creates the new temp table, then it is not a disabling change but if it also cleans up at least 1 existing temp tables, then it must be a disabling change. "Disabling change" is defined by the diff 477e8495c5b6b536b6069eb3640374cb14e9edf6 that introduced YB_DDL_MODE_SILENT_ALTERING:

- before committing DDL transaction with the YB_DDL_MODE_SILENT_ALTERING mode the response cache for particular DB is marked as disabled.
- while response cache on particular DB is disabled any request related to this DB will no use data from the cache. All request will go to a master instead.
- after DDL commit is completed (with any status) response cache for particular DB is marked as enabled

##### Before Transactional DDL (Separate ddl txn)
The code flow upon receiving the `create temp table` command was as follows:
1. Detect the DDL (`YbGetDdlMode`) and start the separate ddl transaction in `YBTxnDdlProcessUtility`. Create temp table starts with `aspects = 0` since it is not a breaking change nor does it require any version increment.
2. Remove temp relations (`RemoveTempRelations`)
    - YBIncrementDdlNestingLevel with mode `YB_DDL_MODE_SILENT_ALTERING`. This is remembered in `ddl_transaction_state.pending`.
    - Remove any temp relations
    - `YBDecrementDdlNestingLevel` which checks if we removed any tables by checking if there was any catalog write. If yes, then it merges the mode `YB_DDL_MODE_SILENT_ALTERING` into ddl_transaction_state.applied and remove from ddl_transaction_state.pending. In other words, it remembers that we've made some changes to the catalog with mode ` YB_DDL_MODE_SILENT_ALTERING`. If no catalog writes were done, then we just remove the mode from `ddl_transaction_state.pending`.
3. Create the temp table
4. Auto commit at the end of the statement

So if we ended up removing any temp relations, the create temp table is correctly treated as a disabling change while if not, then it is treated as a non-disabling change and no cache invalidation is done.

##### With Transactional DDL enabled

1. Detect the DDL (`YbGetDdlMode`) and mark the current txn as executing a ddl in `YBTxnDdlProcessUtility`. Create temp table starts with `aspects = 0` since it is not a disabling change nor does it require any version increment.
2. Remove temp relations (`RemoveTempRelations`)
    - YBAddDdlTxnState with mode `YB_DDL_MODE_SILENT_ALTERING`. This is remembered in `ddl_transaction_state.pending`.
    - Remove any temp relations
    - **There is no call to merge from pending to applied if we did any catalog write.** So no call to cleanup `ddl_transaction_state.pending` as well.
3. Create the temp table
4. Commit at the end of the txn block
    - This checks if there was any catalog write. Since we have done the catalog write of creating the temp table, it always returns true. So, ddl_transaction_state.applied is updated with `YB_DDL_MODE_SILENT_ALTERING` from ddl_transaction_state.pending and the **statement always gets treated as a disabling change regardless of whether we removed any temp tables**.

This is a bug!

To fix this issue, this revision adds the logic to merge the catalog modification aspects from ddl_transaction_state.pending to ddl_transaction_state.applied in every nested ddl step. This way, the correct mode is accounted for at the end of the statement. Function `YBMergeDdlTxnState` has been introduced for it.
Jira: DB-17619, DB-18505

Original commit: 251aeccfbe8355db54cbe7c2bfb5a293e4cfbb20 / D47184

Test Plan:
Enabled transactional DDL and ran the following:

```
./yb_build.sh --cxx-test pgwrapper_pg_catalog_perf-test --gtest_filter PgCatalogPerfTest.SameTempTableCreationWithResponseCache
```

This test fails without the fix since it leads to cache invalidation which is what is asserted in the test. With the fix, it passes.

Reviewers: myang, patnaik.balivada, pjain

Reviewed By: myang

Subscribers: yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D47246