[BACKPORT 2025.2][#28817] docdb: Fix GeoTransactionPromotionTests failing with object locking
Summary:
A handful of GeoTransactionPromotionTests tests:
- GeoTransactionsPromotion.TestOldRegionFailBeforeAbortSentEarly
- GeoTransactionsPromotion.TestOldRegionFailBeforeRPCSend
- GeoTransactionsPromotion.TestOldRegionFailAfterRPCBeforeAbortSentEarly

fail when enable_object_locking_for_table_locks is on. This is because all three tests do the following:
1. initiate a transaction promotion
2. shutdown tserver hosting transaction tablet
3. run select on data
(3) worked fine when enable_object_locking_for_table_locks is off, since no docdb transaction is
used. But with enable_object_locking_for_table_locks on, we now start a docdb transaction for this
read, and this is started as region-local (related #28317). But the region-local transaction tablets
are all on the shutdown tserver, so this gets stuck.

This changes forces the transaction used for those selects to be global.
Jira: DB-18528

Original commit: 4d42274320073182c867a9e34275a2b298732656 / D47234

Test Plan:
With changes rebased over D44416, ran:
```
./yb_build.sh --cxx-test pgwrapper_geo_transactions_promotion-test --gtest_filter GeoTransactionsPromotionTest.TestOldRegionFailBeforeAbortSentEarly
./yb_build.sh --cxx-test pgwrapper_geo_transactions_promotion-test --gtest_filter GeoTransactionsPromotionTest.TestOldRegionFailBeforeRPCSend
./yb_build.sh --cxx-test pgwrapper_geo_transactions_promotion-test --gtest_filter GeoTransactionsPromotionTest.TestOldRegionFailAfterRPCBeforeAbortSentEarly
```

Reviewers: bkolagani, amitanand

Reviewed By: bkolagani

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D47252