[BACKPORT 2025.2][PLAT-18496][CDC][yba] Display the wal_status in replication slot status for CDC metrics
Summary:
Currently we display the value of state present in the stream metadata (obtained by running yb-admin command list_change_data_streams ) in Replication slot status field in the CDC metrics page . This does not provide much debugging information and often leads to confusion regarding expiry status of the stream.

The request is to populate this field using the wal_status field of pg_replication_slots ysql view instead. The wal_status could be either reserved meaning the stream is not yet expired or it could be lost meaning the stream is expired. The pg_replication_slots is a ysql view and can be queried by the slot name as well as the stream id.

Original commit: 9d7689034ce79045672acca18b61253e32201eb4 / D46904

Test Plan:

{F400640}

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: sumukh.phalgaonkar, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D47318