[BACKPORT 2025.1][#28324] YSQL: Use Index Only Scan in INSERT ... ON CONFLICT ... DO NOTHING
Summary:
Unique index constraint checking for `INSERT ... ON CONFLICT` statements is performed by executing an index scan on each of the arbiter indexes.
The scan on the index retrieves the ybctid(s) of the conflicting row(s) from the arbiter index, which is then used to fetch the row itself from the main table.
This results in 2 RPCs per arbiter index: one for the scan on the arbiter index, one for the main table lookup.

However, in case of `INSERT ... ON CONFLICT ... DO NOTHING`, it suffices to know whether the row to be inserted has a conflict in the arbiter index.
The conflicting row in the arbiter index is not used and thus need not be retrieved. Therefore, it suffices to perform an IndexOnlyScan on the arbiter index, thus saving on 1 RPC.

This revision executes IndexOnlyScans to perform unique index constraint checking for `INSERT ... ON CONFLICT ... DO NOTHING` statements.
This optimization is independent of `INSERT ... ON CONFLICT` read batching and is applicable both when read batching is enabled and disabled.

### Locking Considerations
There are two properties in vanilla postgres that make this optimization transactionally correct:
 - Vanilla postgres does not support locking of index tuples. Instead, locks are always acquired on the main table tuple.
 - Index constraint checking (ie. checking arbiter indexes) does not acquire any row locks on the main table.

Therefore, skipping the main table scan in YugabyteDB does not result in any missing lock acquisitions.

To perform lockless unique constraint checks, vanilla postgres relies on two mechanisms:
 - The index scan on the arbiter index is performed using a dirty snapshot (ie. changes made by in-progress transactions are visible)
 - Index tuples in vanilla postgres do not store any transaction visibility information. So, the main table must be looked up to check if the source of conflict is an in-progress transaction.

Therefore, index scans cannot be converted to index-only scans in postgres.
In contrast, YugabyteDB does not suffer from the same problem as a "clean" snapshot is always used by the scan on the arbiter index.

### Future Work
This optimization can be applied to `DO UPDATE` statements when:
 - the old row is not referenced in the SET clause
 - the columns being updated do not have secondary indexes on them (and thus do not require fetching the old row to perform the index update)
 - there are no before-row-insert triggers on the table ([GH-28811](https://github.com/yugabyte/yugabyte-db/issues/28811))
Jira: DB-18006

Original commit: 3de87fb88dc98ce037df601c4beb1b04eed7b2a3 / D46360

Test Plan:
Run the following tests:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch0#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch1#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch2#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch3#schedule'
```

Reviewers: jason

Reviewed By: jason

Subscribers: smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D46386