[#28871] YSQL: Reset transaction read point when required in transactional DDL mode
Summary:
Before this change, if transactional DDL is enabled and a DDL is executed in a read committed isolation transaction, the transaction read point isn't reset for DML statements that are executed after the DDL. This change fixes that.

```
BEGIN ... READ COMMITTED;
DML;
DDL;
DML; --> The read point isn't reset here because YBCPgIsDdlMode() becomes true for the lifetime of the transaction after the first DDL. And so, YbSkipPgSnapshotManagement() stays true and so resetting of the transaction read point is skipped.
```
Jira: DB-18589

Test Plan: ./yb_build.sh release --java-test org.yb.pgsql.TestPgRegressIsolationWithTxnDdl#testPgRegress

Reviewers: stiwary, sanketh

Reviewed By: stiwary

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47341