[#28758] YSQL: only allow internal connection to be relcache init connection
Summary:
Currently, when a new connection needs to rebuild the relcache init file, it sends the
`TriggerRelcacheInitConnection` RPC to the local tserver. The local tserver makes an internal
connection to create the relcache init file. After the RPC returns, the new connection resumes
without having to incur the memory spike that is associated with rebuilding the relcache init file.

The internal connection is made as user "postgres". It is a real user, and therefore can be used to
make an explicit connection. Our C++ unit tests actually connect as "postgres" by default.

If there are concurrent new connections from user "postgres", they will not call
`TriggerRelcacheInitConnection` RPC. As a result they all pay the cost of rebuilding the relcache
init file, leading to memory spike in all of them.

In this diff, I made a change to only allow internally generated connection to rebuild relcache init
file. I think internally connection is always made as "postgres". If the superuser "postgres" makes a
connection externally (e.g., from ysqlsh), it will also trigger a call to
`TriggerRelcacheInitConnection` RPC.

I found that existing MiniCluster based C++ unit tests do not have separate
`pgsql_proxy_bind_address` set for each tserver. For example, if the test cluster has RF-3 and
therefore has 3 tservers, all of them share the same `pgsql_proxy_bind_address` (which tserver-1 is
created with and is used by postmaster of tserver-1). With this diff if an external postgres connection
on node-2 triggers an internal postgres connection, the internal connection is made from tserver-2 and
this internal connection failed to authenticate because it used the `pg_auth_key` of tserver-2
(via `GetSharedMemoryPostgresAuthKey`) and then talked to postmaster of tserver-1 via the common
`pgsql_proxy_bind_address`, which is expecting the `pg_auth_key` from tserver-1.

I made changes to MiniCluster based C++ unit tests so that each tserver uses a unique
`pgsql_proxy_bind_address` that is shared between its postmaster and itself. In this way an internal
"postgres" connection made from tserver-2 will uses the unique `pgsql_proxy_bind_address` of
tserver-2 talking to the postmaster of tserver-2.

For XCluster based tests, there is only one postmaster generated in the producer cluster, all 3
tservers in the producer cluster share the same `pg_auth_key` via the use of
`FLAGS_TEST_pg_auth_key`. Same is true for the consumer cluster. I kept it that way for XCluster
based tests.

Some unit tests are updated (e.g., metrics numbers) because a first connection in the test is now no
longer the first connection. It will trigger an internal connection which is now the first
connection.
Jira: DB-18472

Test Plan: jenkins

Reviewers: kfranz, sanketh

Reviewed By: kfranz

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D47358