[#28889] xClusterDDLRepl: Fix IncrementalSafeTimeBumpWithDdlQueueStepdowns in ASAN
Summary:
After introducing advisory locks to ddl_queue_handler, only one handler can be active at once. This test creates a handler, pauses (which deletes the handler), then resumes (which recreates the handler). When the handler is deleted, it takes a bit for the pg session to shutdown, which means that the advisory lock is held on for an extended period of time. During this period the next handler can't grab the advisory lock, so it can't make progress. In build types that are more time sensitive (eg ASAN) , this causes checks to occur before the new handler is ready.

Fix is to try to release all advisory locks when shutting down the ddl queue handler.
Jira: DB-18611

Test Plan:
```
ybd asan --cxx-test integration-tests_xcluster_ddl_replication-test --gtest_filter XClusterDDLReplicationTest.IncrementalSafeTimeBumpWithDdlQueueStepdowns -n 40
```

Reviewers: hsunder, xCluster

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D47362